+++ leaderboardwidget.cpp
@@ -8,10 +8,10 @@ LeaderBoardWidget::LeaderBoardWidget(QWidget *parent) : QWidget(parent)
 {
     playerCreationWidget->hide();
     addBtn->setText("Add new Player");
-    for(int i = 0; i<numOfPlayers; i++)
+    for(unsigned int i = 0; i<numOfPlayers; i++)
         lay->addWidget(players.at(i), i, 0);
     lay->addWidget(addBtn, numOfPlayers, 0);
-    for(int i = 0; i<numOfPlayers; i++)
+    for(unsigned int i = 0; i<numOfPlayers; i++)
     {
         players.at(i)->accept->setEnabled(i == currentPlayer?true:false);
         players.at(i)->lSpinBox->setEnabled(i == currentPlayer?true:false);
@@ -30,8 +30,8 @@ void LeaderBoardWidget::sortByBidding()
 { //Sort by Bidding, Person with the lowest bid at the top, might not be a good idea
     PlayerBiddingWidget * sortedPlayers;
     sortedPlayers = (PlayerBiddingWidget*) calloc(numOfPlayers, sizeof(PlayerBiddingWidget*));
-    u_int16_t minBid = players.at(0)->getBidding();
-    for(u_int8_t i = 0; i<numOfPlayers-1; i++)
+    unsigned int minBid = players.at(0)->getBidding();
+    for(unsigned int i = 0; i<numOfPlayers-1; i++)
     {
         //Do Stuff, this has to be done better
         //lay->addWidget(lay->takeAt(0), i, 0);
@@ -49,10 +49,10 @@ void LeaderBoardWidget::addPlayer(User * newUser)
     //players[numOfPlayers] = player;
     lay->addWidget(players.at(numOfPlayers));
 
-    for(int i = 0; i<numOfPlayers; i++)
+    for(unsigned int i = 0; i<numOfPlayers; i++)
         lay->addWidget(players.at(i), i, 0);
     lay->addWidget(addBtn, numOfPlayers, 0); //New Player Button under all Players
-    for(int i = 0; i<numOfPlayers; i++)
+    for(unsigned int i = 0; i<numOfPlayers; i++)
     {
         players.at(i)->accept->setEnabled(i == currentPlayer?true:false);
         players.at(i)->lSpinBox->setEnabled(i == currentPlayer?true:false);
